pyproject.toml
[tool.poetry]
name = "fake"
version = "0.1.0"
description = ""
authors = ["Alym Abdullayev <alymzhan.abdullayev@daktela.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.1.4"
djangorestframework = "^3.14.0"
django-environ = "^0.9.0"
django-shell-plus = "^1.1.7"
django-guid = "^3.3.0"
django-channels = "^0.7.0"
celery = "^5.2.7"
psycopg2 = "^2.9.5"
Faker = "^15.3.4"
markdown = "^3.4.1"
django-filter = "^22.1"
pyyaml = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
